server/requirements.txt
websockets==11.0.3
uuid==1.30
google-generativeai==0.3.2
python-dotenv==1.0.0
